image-resize/img-resize-lambda/requirements.txt
Pillow==7.2.0
boto3==1.14.28